requirements.txt
aiohttp==3.9.5
peewee==3.17.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9